Custom script for Ensembl VEP
This script generates the annotation for the variants retrieved from gnomADv3.1.2
It also generates a reference.fa and mutated.fa file for each chromosome. These can then be used in downstream analyses such as protein annotation with InterPro, clustering based on sequence identity and coverage with CD-HIT and so on.